Fix: Core-SG weights corretos
